Internal editor now compiles and run in PC x64

+++ code/editor/OsirisStatusDlg.cpp
@@ -78,7 +78,7 @@ void COsirisStatusDlg::SetScript(char *script, tD3XProgram *d3x)
 	m_Script=script;
 }
 
-int COsirisStatusDlg::DoModal() 
+intptr_t COsirisStatusDlg::DoModal() 
 {
 	// TODO: Add your specialized code here and/or call the base class
