+++ code/editor/OsirisStatusDlg.h
@@ -51,7 +51,7 @@ public:
 	// ClassWizard generated virtual function overrides
 	//{{AFX_VIRTUAL(COsirisStatusDlg)
 	public:
-	virtual int DoModal();
+	virtual intptr_t DoModal();
 	protected:
 	virtual void DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
 	//}}AFX_VIRTUAL
